requirements.txt
brotlicffi==1.1.0.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
idna==3.7
pycparser==2.22
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.2
